requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0

# gRPC
grpcio==1.59.3
grpcio-tools==1.59.3
grpcio-reflection==1.59.3
protobuf==4.25.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
cryptography==41.0.7

# Caching & Performance
redis==5.0.1
django-redis==5.4.0
hiredis==2.2.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Monitoring & Logging
django-prometheus==2.3.1
sentry-sdk==1.39.1

# API Documentation
drf-spectacular==0.27.0
drf-spectacular-sidecar==2023.12.1

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
faker==20.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
pylint==3.0.3

# Environment & Configuration
python-dotenv==1.0.0
environs==10.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
